Split flow classes to files
